refactor: consolidate and enhance tool calling functions
- Removed deprecated `generateTextOllama` and `streamTextOllama` functions, integrating their functionality into `generateText` and `streamText` for improved reliability and simplicity.
- Updated README and example files to reflect changes in function usage and enhanced tool calling capabilities.
- Ensured all examples demonstrate the new unified approach for generating text and streaming responses with guaranteed reliability.
