+++ .changeset/petite-rocks-cover.md
@@ -1,0 +1,29 @@
+---
+'ai-sdk-ollama': patch
+---
+
+## Enhanced Function Renaming & Documentation Improvements
+
+### Function Renaming
+- Renamed `generateTextOllama` to `generateText` (enhanced version from ai-sdk-ollama)
+- Renamed `streamTextOllama` to `streamText` (enhanced version from ai-sdk-ollama)
+- Maintains backward compatibility while providing clearer API
+
+### Documentation Improvements
+- **README.md**: Complete restructure with better user flow
+  - Added Quick Start section with immediate installation and basic example
+  - Moved value proposition ("Why Choose") section earlier
+  - Added dedicated "Enhanced Tool Calling" section highlighting main differentiator
+  - Reorganized examples under "More Examples" for better progression
+  - Removed redundant content and improved clarity
+
+- **packages/ai-sdk-ollama/README.md**: Applied same improvements
+  - Consistent structure with main README
+  - Better user journey from basic to advanced features
+  - Updated table of contents to match new structure
+
+### Key Benefits
+- **Better Developer Experience**: Clearer function names and improved documentation flow
+- **Enhanced Tool Calling**: Highlighted the main selling point with dedicated section
+- **User-Friendly**: Users can now get started in 30 seconds and understand value immediately
+- **Consistent**: Both READMEs now have the same improved structure and flow
